backend/requirements.txt
# PolicyLens AI — Python dependencies
# Install: pip install -r requirements.txt

# API Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0          # SQLite async (dev); swap for asyncpg for PostgreSQL
asyncpg==0.29.0            # PostgreSQL async driver for SQLAlchemy async engine
psycopg2-binary==2.9.9     # PostgreSQL (production)
pydantic==1.10.26

# Document Parsing
pymupdf==1.23.8            # PyMuPDF — fitz
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0

# NLP & ML
spacy==3.7.2
sentence-transformers==2.7.0  # Updated for torch compatibility
faiss-cpu==1.13.2
transformers==4.37.2
torch==2.11.0                 # Updated to available version (2.1.2 no longer available)
aif360==0.6.1
textstat==0.7.3

# LangChain for text chunking
langchain==0.1.5

# Translation
argostranslate==1.9.6      # Offline fallback translation

# Utilities
python-multipart==0.0.6    # File uploads
python-dotenv==1.0.0
